Move commands into its own package
