+++ src/main/java/jank/JankBotException.java
@@ -4,7 +4,7 @@ package jank;
  * Custom exception class for the chatbot
  */
 public class JankBotException extends Exception {
-    JankBotException(String msg) {
+    public JankBotException(String msg) {
         super(msg);
     }
 }
