+++ src/main/java/jank/command/Command.java
@@ -1,0 +1,4 @@
+package jank.command;
+
+interface Command {
+}
